+++ .gitignore
@@ -35,3 +35,4 @@ dev-dist
 *storybook.log
 
 .env
+.pem
