+++ localhost-key.pem
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQC10SiiwKMvoaGB
+Jm4HWYj40SP99zPWeB/5PA6pC8lQEK+YRdAt2mipWZ9g93wZPQ8kI/94CpT4wElo
+YK4w2xCuuPnEowW6nGMkrwFs1hJxnqEANPJL0MenKLF1jNMTxR2RH4UPePZlyiHX
+/wBSl6NwoCJK0nrCZeuymCmF6p3PZkqVYmRk7WFbLBoaGynR6gls4Zz238Sluik1
+R8zdMGLeZ4fFx9yW10BokO2AkgIgzgACO9kHIO9hUcKuuSqvoquEkd/WmD2VLhAC
+FYyEep34toNkttDocYj/wXEAZAUSMmZ1nSMQPD1ef+XJMsNzkmEW1rhDQRFMgHUs
+taVILWnhAgMBAAECggEBAJFlfyKP/K/liJsYWVsFTdT0otzAqDK1g5QzY+HprAac
+IWAbTDyu8Y2kPyEzKImiW45iPzUG+VTnqIyQ7Bvjw7aF9C880yMFVv+J+BDKLJux
+6DybGrxobJEPqMJBaLbhpiBc3XKfICQMff4WTXK/ExPI60N8iHEXQmJi7qnKEoM8
+h4tIOq22Z4f6HNXg1uWoQbA0+Du0lNWaGFHHieZkeMVBpP0noSdM1x6QeRbNV234
+midoYt9tyYA7doFHpfUYCK8vIBlJLCamJrQGsL1+pWVqfdr8q0EgRugXT5ZV3Dq2
+ox1kSA3nykUQ4P+dbRYADqAQNCvIMY2rPG9VDJ2f83UCgYEAzEO5lQyegrJ6/uYn
+uwNfIdH/tfiTt/dhj7+iP97dajCmxmVQmMFJGI6H0CnHv4FOz8wbhYk2wAwtLCQ+
+CZhVF47DbP2EuZIEQXG3LbsnQrDvbwQsemfgH+/QpL1//UtO6BhHanKMK+NbVBHe
+YHKuhZVciRV+G+Slaky3DldOGxMCgYEA4932aGVtiNGqg0zGvURB19wH4N1cWNT6
+rDoeOQ+Lq/Ts3hW+BIoGHGyszRjwPgFo3WWbfcx8dM/V/sVhIcfxKo2X3HeF/sjB
+iEraaFIEFsDTLDT++PoIPUJ+6ecy2zEcbhV5h0Uu9MeMRSlPoCrFOkM84XWD8KTR
+Mwsq6IprMbsCgYARlbsCU+XJhqgEDbhZ5z5nYPbMqau96LQjdfH34a3aGHojFlHL
+9MfgL1UN3Krvl6g7R20PH/LwsZup8yzMEqswrf+xAFpOx31KTQniC/yjRyw1uHxD
+xXGMDk89/3pnSfZvELgFN/tNQ/N211r3nqfbpDIlMw56RELhVd1khKuatQKBgGe5
+Odj7koidKAjsgerVj0M7HTr2iHrHN/P94JQIx4NW0hulugFG8tEthts1ZuZSrfFZ
+txXkZyKlGLwIDNbbtKOyIFZy8cOwjqgDqreh3cqdq4KL/SRIbEe2ZR8D5tpv2Ikj
+rVYt39n8m5U/4Mb/sRd8U9ItNAnufGvHwZhoqbLRAoGBAIkNOfptAYM0jsUmXz97
+lG03LGmp3hLyjHimPCHv47qSNauLefKmZ6JoHxUPSnceCmvCbUWIcQVgdzKvahc6
+AGmLSnM/QDfvvEWlFkKIecai6B945Au0Tl+FpZBfwC4LXQUhBqwFCOaJ257gCy/h
+yg8ygmW2B6PcH4BZhvSfV4vz
+-----END PRIVATE KEY-----
