+++ localhost.pem
@@ -1,0 +1,26 @@
+-----BEGIN CERTIFICATE-----
+MIIEczCCAtugAwIBAgIQdaIh4bJ1LnwOhRfC6AainTANBgkqhkiG9w0BAQsFADCB
+ozEeMBwGA1UEChMVbWtjZXJ0IGRldmVsb3BtZW50IENBMTwwOgYDVQQLDDNhbmpv
+bmctaW5AYW5qb25nLWluLXVpLU1hY0Jvb2tQcm8ubG9jYWwgKOyViOyiheyduCkx
+QzBBBgNVBAMMOm1rY2VydCBhbmpvbmctaW5AYW5qb25nLWluLXVpLU1hY0Jvb2tQ
+cm8ubG9jYWwgKOyViOyiheyduCkwHhcNMjUwMTA0MDkwOTU0WhcNMjcwNDA0MDkw
+OTU0WjBnMScwJQYDVQQKEx5ta2NlcnQgZGV2ZWxvcG1lbnQgY2VydGlmaWNhdGUx
+PDA6BgNVBAsMM2Fuam9uZy1pbkBhbmpvbmctaW4tdWktTWFjQm9va1Byby5sb2Nh
+bCAo7JWI7KKF7J24KTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBALXR
+KKLAoy+hoYEmbgdZiPjRI/33M9Z4H/k8DqkLyVAQr5hF0C3aaKlZn2D3fBk9DyQj
+/3gKlPjASWhgrjDbEK64+cSjBbqcYySvAWzWEnGeoQA08kvQx6cosXWM0xPFHZEf
+hQ949mXKIdf/AFKXo3CgIkrSesJl67KYKYXqnc9mSpViZGTtYVssGhobKdHqCWzh
+nPbfxKW6KTVHzN0wYt5nh8XH3JbXQGiQ7YCSAiDOAAI72Qcg72FRwq65Kq+iq4SR
+39aYPZUuEAIVjIR6nfi2g2S20OhxiP/BcQBkBRIyZnWdIxA8PV5/5ckyw3OSYRbW
+uENBEUyAdSy1pUgtaeECAwEAAaNeMFwwDgYDVR0PAQH/BAQDAgWgMBMGA1UdJQQM
+MAoGCCsGAQUFBwMBMB8GA1UdIwQYMBaAFK8f0Cygzni2ugrz4yclmes4Nn4nMBQG
+A1UdEQQNMAuCCWxvY2FsaG9zdDANBgkqhkiG9w0BAQsFAAOCAYEAw2uPc8wJ0bYY
+f3wzQgqaccj7xdvXRa+k+J7IJ3M21XyvjjleV8w+n9FjTr47SM7/VA6qRok8PN4b
+SPH4IXJiNMVxEuUYX1z6YpyvuulnUsGDXvvTAExcdYE3msfr6UE0pU+rFfNVxUb3
+7bmbMFD+R43F5hgSSl/R9zoQG43Uc1GZ8jhGZHuScZB3PE9fuhfz3OszEItYIbMz
+lVDIMbKFhrVJva5/pCcY2PTMQuHjxTcYITg+HiA4rmETPwK/sB3Sfn5WLj0WtJyn
+dKqkZsP8EBwqUgbQcwkSWWvQ3P2s9anv9VKqVVNVww6aWzS2VIX3D3P2CCo7rthH
++ysJdDUXxV0XOSXFv8n8a8BTCSIAnd5dIfPlPm1qXHesGL8PTcDRpKISQpXmb5Wq
+1pbyv5dGB59LmLGi5e2bWZJVqe2AwmUvWw/d43xwu84EzUineKc+phC7PgyNM0Lx
+mF+GOn1bNv5ZrNLiFaNqc0l07QZUGvkF+NuRJMgnLtWeUXMbhOHQ
+-----END CERTIFICATE-----
