+++ package.json
@@ -32,7 +32,7 @@
     "@types/react-icons": "^3.0.0",
     "axios": "^1.7.8",
     "chakra-ui-bottom-navigation": "^2.0.1",
-    "firebase": "^11.0.2",
+    "firebase": "8.10.0",
     "framer-motion": "^6",
     "glob": "^11.0.0",
     "js-cookie": "^3.0.5",
