+++ src/main.tsx
@@ -8,8 +8,7 @@ import { ThemeProvider } from '@emotion/react';
 import { GlobalStyles } from '@/styles/GlobalStyles';
 import theme from '@/styles/theme';
 import { BottomNavigationStyleConfig as BottomNavigation } from 'chakra-ui-bottom-navigation';
-import { initializeApp } from 'firebase/app';
-import { getMessaging, getToken, onMessage } from 'firebase/messaging';
+import './firesbase-message';
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -27,64 +26,21 @@ const chakraTheme = extendTheme({
   },
 });
 
-// Firebase 초기화
-const firebaseConfig = {
-  apiKey: 'AIzaSyCFC3QJKAZhz1R0k-h58wJA8_Rb_PbyiL4',
-  authDomain: 'q-feed.firebaseapp.com',
-  projectId: 'q-feed',
-  storageBucket: 'q-feed.firebasestorage.app',
-  messagingSenderId: '804246377517',
-  appId: '1:804246377517:web:71270af160949939da14a4',
-};
-
-const firebaseApp = initializeApp(firebaseConfig);
-const messaging = getMessaging(firebaseApp);
-
-// FCM 초기화 및 토큰 요청
-const requestFCMToken = async () => {
-  try {
-    // 알림 권한 요청
-    const permission = await Notification.requestPermission();
-    if (permission !== 'granted') {
-      console.warn('알림 권한이 허용되지 않았습니다.');
-      return;
-    }
-
-    // FCM 토큰 요청
-    const token = await getToken(messaging, {
-      vapidKey:
-        'BKvBPha3ZSEI7Xb55-iWciONGqfKYtYgdj6kGWVe-mZDoeKYCCGwmAJaA12wl3zllzU5LCGX4Ar3_8Fix2QqEQ8',
-    });
-
-    if (token) {
-      console.log('FCM 토큰:', token);
-    } else {
-      console.warn('FCM 토큰을 가져올 수 없습니다. 알림 권한이 허용되지 않았을 수 있습니다.');
-    }
-  } catch (error) {
-    console.error('FCM 토큰 가져오기 실패:', error);
-  }
-};
-
-// FCM 포그라운드 메시지 처리
-onMessage(messaging, (payload) => {
-  console.log('포그라운드 메시지 수신:', payload);
-  // TODO: 알림 UI 표시
-});
-
-// Service Worker 등록
 if ('serviceWorker' in navigator) {
   navigator.serviceWorker
     .register('/firebase-messaging-sw.js')
     .then((registration) => {
       console.log('Service Worker 등록 성공:', registration);
-      requestFCMToken(); // Service Worker가 등록된 후 토큰 요청
+
+      // Firebase Messaging 초기화
+      import('./firesbase-message').then(({ requestPermission }) => {
+        requestPermission();
+      });
     })
     .catch((error) => {
       console.error('Service Worker 등록 실패:', error);
     });
 }
-
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
